util/cbfstool/fmd_parser: Add rules to define extension regions
Define a new EXT flag that tags certain regions in the flash chip to be
extended to an alternate storage (e.g. disk). This new EXT flag applies
only to regions with coreboot filesystem (CBFS). The extension regions
inherit the name of the regions that are getting extended with an
"ECBFS_" prefix. Hence block "ECBFS_" prefix for any regions in the
flash chips. Add rules in fmd_parser to define these extension regions.

BUG=b:308868205, b:308862651
BRANCH=none
TEST=flex -L --outfile=fmd_scanner.c_shipped \
        --header-file=fmd_scanner.h_shipped fmd_scanner.l
     bison -d fmd_parser.y -l --header=fmd_parser.h_shipped \
        --output=fmd_parser.c_shipped

Change-Id: I9fb4e555b5d2dd5316804f6c13755270db60cb60
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/91803
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Jakub "Kuba" Czapiga <czapiga@google.com>